+++ src/App.css
@@ -152,8 +152,17 @@ body {
 
 .modal__button {
   cursor: pointer;
-  height: -moz-min-content;
-  height: -webkit-min-content;
   height: min-content;
-  width: fit-content;
+  margin-top: 18px;
+  padding: 10px;
+  padding-right: 15px;
+  padding-left: 15px;
+  border: none;
+  border-radius: 100px;
+  font-size: 14px;
+  background-color: var(--primary);
+}
+
+.modal__button:hover {
+  background-color: #c0c8d2;
 }
