+++ src/components/NavBar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./styles/App.css";
+import "../App.css"
 import InformationCircleIcon from "@heroicons/react/24/outline/InformationCircleIcon";
 
 const NavBar = ({ showInstructions, setShowInstructions }) => {
