--- src/styles/Keyboard.css
@@ -1,24 +1,0 @@
-.keyboard {
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    justify-content: center;
-    gap: 10px;
-  }
-  .keyboard__row {
-    display: flex;
-    justify-content: center;
-    gap: 6px;
-  }
-  
-  .keyboard__letter {
-    width: 50px;
-    height: 50px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-transform: capitalize;
-    background: var(--primary);
-    border-radius: 7px;
-    cursor: pointer;
-  }
