--- src/styles/Modal.css
@@ -1,74 +1,0 @@
-.backdrop {
-  position: absolute;
-  left: 0;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(150, 150, 150, 0.7);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.modal {
-  width: 400px;
-  height: fit-content;
-  background: white;
-  padding: 10px;
-  border-radius: 5px;
-  text-align: center;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  flex-direction: column;
-  box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.3);
-}
-
-.modal__title {
-  font-family: var(--font-family);
-  font-size: 22px;
-}
-
-.modal__button {
-  cursor: pointer;
-  height: -moz-min-content;
-  height: -webkit-min-content;
-  height: min-content;
-  width: fit-content;
-}
-.nav-bar {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  background-color: var(--primary);
-  color: black;
-  box-shadow: 0px 4px 6px -1px rgba(0, 0, 0, 0.1),
-    0px 2px 4px -1px rgba(0, 0, 0, 0.2);
-}
-
-.nav-title {
-  font-family: var(--font-family);
-  font-weight: bold;
-  font-size: 1.5em;
-  align-self: center;
-}
-
-.info-icon {
-  height: 30px;
-  width: 30px;
-  cursor: pointer;
-  align-self: start;
-}
-
-.custom-h {
-  height: 100vh;
-}
-
-.exit-button-container {
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-}
-
-.exit-button {
-  cursor: pointer;
-}
