--- src/styles/NavBar.css
@@ -1,37 +1,0 @@
-.nav-bar {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  background-color: var(--primary); /* Use the primary color variable */
-  color: black; /* Assuming the text color for the app is black */
-  box-shadow: 0px 4px 6px -1px rgba(0, 0, 0, 0.1),
-    0px 2px 4px -1px rgba(0, 0, 0, 0.2);
-}
-
-.nav-title {
-  font-family: var(--font-family);
-  font-weight: bold;
-  font-size: 1.5em;
-  align-self: center;
-}
-
-.info-icon {
-  height: 30px;
-  width: 30px;
-  cursor: pointer;
-  align-self: start;
-}
-
-.custom-h {
-  height: 100vh;
-}
-
-.exit-button-container {
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-}
-
-.exit-button {
-  cursor: pointer;
-}
